add post request example with textfield
